refactored frontend public/explore API
