+++ frontend/lib/api/client-public.ts
@@ -5,13 +5,20 @@ import { ApiRequestInstance } from "~/lib/api/types/non-generated";
 
 export class PublicApi {
   public validators: ValidatorsApi;
-  public explore: ExploreApi;
   public shared: SharedApi;
 
   constructor(requests: ApiRequestInstance) {
     this.validators = new ValidatorsApi(requests);
-    this.explore = new ExploreApi(requests);
     this.shared = new SharedApi(requests);
+  }
+}
+
+export class PublicExploreApi extends PublicApi {
+  public explore: ExploreApi;
+
+  constructor(requests: ApiRequestInstance, groupSlug: string) {
+    super(requests);
+    this.explore = new ExploreApi(requests, groupSlug);
 
     Object.freeze(this);
   }
